게시물 수정, 삭제시 권한체크

+++ src/main/java/com/jhs/exam/exam2/http/controller/UsrArticleController.java
@@ -58,6 +58,13 @@ public class UsrArticleController extends Controller {
 			return;
 		}
 
+		ResultData actorCanDeleteRd = articleService.actorCanDelete(rq.getLoginedMember(), article);
+
+		if (actorCanDeleteRd.isFail()) {
+			rq.historyBack(actorCanDeleteRd.getMsg());
+			return;
+		}
+
 		articleService.delete(id);
 
 		rq.replace(Ut.f("%d번 게시물을 삭제하였습니다.", id), redirectUri);
@@ -139,6 +146,20 @@ public class UsrArticleController extends Controller {
 			return;
 		}
 
+		Article article = articleService.getForPrintArticleById(id);
+
+		if (article == null) {
+			rq.historyBack(Ut.f("%d번 게시물이 존재하지 않습니다.", id));
+			return;
+		}
+
+		ResultData actorCanModifyRd = articleService.actorCanModify(rq.getLoginedMember(), article);
+
+		if (actorCanModifyRd.isFail()) {
+			rq.historyBack(actorCanModifyRd.getMsg());
+			return;
+		}
+
 		ResultData modifyRd = articleService.modify(id, title, body);
 
 		rq.replace(modifyRd.getMsg(), redirectUri);
@@ -153,6 +174,13 @@ public class UsrArticleController extends Controller {
 		}
 
 		Article article = articleService.getForPrintArticleById(id);
+
+		ResultData actorCanModifyRd = articleService.actorCanModify(rq.getLoginedMember(), article);
+
+		if (actorCanModifyRd.isFail()) {
+			rq.historyBack(actorCanModifyRd.getMsg());
+			return;
+		}
 
 		if (article == null) {
 			rq.historyBack(Ut.f("%d번 게시물이 존재하지 않습니다.", id));
